+++ src/js/utils/constant.js
@@ -149,12 +149,7 @@ const winStatics =(winNumber,earning)=> ` <div class="modal">
         </div>
       </div>`
 
-const alertMessage ={
-  winningOverNum : '당첨번호 값은 1이상 45이하이어야한다.',
-  bonusOverNum : '보너스번호 값은 1이상 45이하이어야한다.',
-  duplicationNum : '중복된 숫자는 불가능하다.',
-  numInTheThousands : '천의자리만 가능하다'
-}
 
 
-export {$,$All,app,lotteImage,appBuySection,inputLastWeekNumber,alertMessage,winStatics}
+
+export {$,$All,app,lotteImage,appBuySection,inputLastWeekNumber,winStatics}
